@@ -1,23 +1,23 @@
+from __future__ import annotations
+
 import typing as t
 from threading import local
 
 if t.TYPE_CHECKING:
-    import typing_extensions as te
-
     from .core import Context
 
 _local = local()
 
 
 @t.overload
-def get_current_context(silent: "te.Literal[False]" = False) -> "Context": ...
+def get_current_context(silent: t.Literal[False] = False) -> Context: ...
 
 
 @t.overload
-def get_current_context(silent: bool = ...) -> t.Optional["Context"]: ...
+def get_current_context(silent: bool = ...) -> Context | None: ...
 
 
-def get_current_context(silent: bool = False) -> t.Optional["Context"]:
+def get_current_context(silent: bool = False) -> Context | None:
     """Returns the current click context.  This can be used as a way to
     access the current context object from anywhere.  This is a more implicit
     alternative to the :func:`pass_context` decorator.  This function is
@@ -41,7 +41,7 @@ def get_current_context(silent: bool = False) -> t.Optional["Context"]:
     return None
 
 
-def push_context(ctx: "Context") -> None:
+def push_context(ctx: Context) -> None:
     """Pushes a new context to the current stack."""
     _local.__dict__.setdefault("stack", []).append(ctx)
 
@@ -51,7 +51,7 @@ def pop_context() -> None:
     _local.stack.pop()
 
 
-def resolve_color_default(color: t.Optional[bool] = None) -> t.Optional[bool]:
+def resolve_color_default(color: bool | None = None) -> bool | None:
     """Internal helper to get the default value of the color flag.  If a
     value is passed it's returned unchanged, otherwise it's looked up from
     the current context.
