@@ -22,6 +22,9 @@ Copyright 2002-2006 Python Software Foundation. All rights reserved.
 # maintained by the Python Software Foundation.
 # Copyright 2001-2006 Gregory P. Ward
 # Copyright 2002-2006 Python Software Foundation
+from __future__ import annotations
+
+import collections.abc as cabc
 import typing as t
 from collections import deque
 from gettext import gettext as _
@@ -33,8 +36,6 @@ from .exceptions import NoSuchOption
 from .exceptions import UsageError
 
 if t.TYPE_CHECKING:
-    import typing_extensions as te
-
     from .core import Argument as CoreArgument
     from .core import Context
     from .core import Option as CoreOption
@@ -49,8 +50,8 @@ _flag_needs_value = object()
 
 
 def _unpack_args(
-    args: t.Sequence[str], nargs_spec: t.Sequence[int]
-) -> t.Tuple[t.Sequence[t.Union[str, t.Sequence[t.Optional[str]], None]], t.List[str]]:
+    args: cabc.Sequence[str], nargs_spec: cabc.Sequence[int]
+) -> tuple[cabc.Sequence[str | cabc.Sequence[str | None] | None], list[str]]:
     """Given an iterable of arguments and an iterable of nargs specifications,
     it returns a tuple with all the unpacked arguments at the first index
     and all remaining arguments as the second.
@@ -62,10 +63,10 @@ def _unpack_args(
     """
     args = deque(args)
     nargs_spec = deque(nargs_spec)
-    rv: t.List[t.Union[str, t.Tuple[t.Optional[str], ...], None]] = []
-    spos: t.Optional[int] = None
+    rv: list[str | tuple[str | None, ...] | None] = []
+    spos: int | None = None
 
-    def _fetch(c: "te.Deque[V]") -> t.Optional[V]:
+    def _fetch(c: deque[V]) -> V | None:
         try:
             if spos is None:
                 return c.popleft()
@@ -108,7 +109,7 @@ def _unpack_args(
     return tuple(rv), list(args)
 
 
-def split_opt(opt: str) -> t.Tuple[str, str]:
+def _split_opt(opt: str) -> tuple[str, str]:
     first = opt[:1]
     if first.isalnum():
         return "", opt
@@ -117,63 +118,29 @@ def split_opt(opt: str) -> t.Tuple[str, str]:
     return first, opt[1:]
 
 
-def normalize_opt(opt: str, ctx: t.Optional["Context"]) -> str:
+def _normalize_opt(opt: str, ctx: Context | None) -> str:
     if ctx is None or ctx.token_normalize_func is None:
         return opt
-    prefix, opt = split_opt(opt)
+    prefix, opt = _split_opt(opt)
     return f"{prefix}{ctx.token_normalize_func(opt)}"
 
 
-def split_arg_string(string: str) -> t.List[str]:
-    """Split an argument string as with :func:`shlex.split`, but don't
-    fail if the string is incomplete. Ignores a missing closing quote or
-    incomplete escape sequence and uses the partial token as-is.
-
-    .. code-block:: python
-
-        split_arg_string("example 'my file")
-        ["example", "my file"]
-
-        split_arg_string("example my\\")
-        ["example", "my"]
-
-    :param string: String to split.
-    """
-    import shlex
-
-    lex = shlex.shlex(string, posix=True)
-    lex.whitespace_split = True
-    lex.commenters = ""
-    out = []
-
-    try:
-        for token in lex:
-            out.append(token)
-    except ValueError:
-        # Raised when end-of-string is reached in an invalid state. Use
-        # the partial token as-is. The quote or escape character is in
-        # lex.state, not lex.token.
-        out.append(lex.token)
-
-    return out
-
-
-class Option:
+class _Option:
     def __init__(
         self,
-        obj: "CoreOption",
-        opts: t.Sequence[str],
-        dest: t.Optional[str],
-        action: t.Optional[str] = None,
+        obj: CoreOption,
+        opts: cabc.Sequence[str],
+        dest: str | None,
+        action: str | None = None,
         nargs: int = 1,
-        const: t.Optional[t.Any] = None,
+        const: t.Any | None = None,
     ):
         self._short_opts = []
         self._long_opts = []
-        self.prefixes: t.Set[str] = set()
+        self.prefixes: set[str] = set()
 
         for opt in opts:
-            prefix, value = split_opt(opt)
+            prefix, value = _split_opt(opt)
             if not prefix:
                 raise ValueError(f"Invalid start character for option ({opt})")
             self.prefixes.add(prefix[0])
@@ -196,7 +163,7 @@ class Option:
     def takes_value(self) -> bool:
         return self.action in ("store", "append")
 
-    def process(self, value: t.Any, state: "ParsingState") -> None:
+    def process(self, value: t.Any, state: _ParsingState) -> None:
         if self.action == "store":
             state.opts[self.dest] = value  # type: ignore
         elif self.action == "store_const":
@@ -212,16 +179,16 @@ class Option:
         state.order.append(self.obj)
 
 
-class Argument:
-    def __init__(self, obj: "CoreArgument", dest: t.Optional[str], nargs: int = 1):
+class _Argument:
+    def __init__(self, obj: CoreArgument, dest: str | None, nargs: int = 1):
         self.dest = dest
         self.nargs = nargs
         self.obj = obj
 
     def process(
         self,
-        value: t.Union[t.Optional[str], t.Sequence[t.Optional[str]]],
-        state: "ParsingState",
+        value: str | cabc.Sequence[str | None] | None,
+        state: _ParsingState,
     ) -> None:
         if self.nargs > 1:
             assert value is not None
@@ -244,15 +211,15 @@ class Argument:
         state.order.append(self.obj)
 
 
-class ParsingState:
-    def __init__(self, rargs: t.List[str]) -> None:
-        self.opts: t.Dict[str, t.Any] = {}
-        self.largs: t.List[str] = []
+class _ParsingState:
+    def __init__(self, rargs: list[str]) -> None:
+        self.opts: dict[str, t.Any] = {}
+        self.largs: list[str] = []
         self.rargs = rargs
-        self.order: t.List[CoreParameter] = []
+        self.order: list[CoreParameter] = []
 
 
-class OptionParser:
+class _OptionParser:
     """The option parser is an internal class that is ultimately used to
     parse options and arguments.  It's modelled after optparse and brings
     a similar but vastly simplified API.  It should generally not be used
@@ -264,9 +231,12 @@ class OptionParser:
 
     :param ctx: optionally the :class:`~click.Context` where this parser
                 should go with.
+
+    .. deprecated:: 8.2
+        Will be removed in Click 9.0.
     """
 
-    def __init__(self, ctx: t.Optional["Context"] = None) -> None:
+    def __init__(self, ctx: Context | None = None) -> None:
         #: The :class:`~click.Context` for this parser.  This might be
         #: `None` for some advanced use cases.
         self.ctx = ctx
@@ -285,19 +255,19 @@ class OptionParser:
             self.allow_interspersed_args = ctx.allow_interspersed_args
             self.ignore_unknown_options = ctx.ignore_unknown_options
 
-        self._short_opt: t.Dict[str, Option] = {}
-        self._long_opt: t.Dict[str, Option] = {}
+        self._short_opt: dict[str, _Option] = {}
+        self._long_opt: dict[str, _Option] = {}
         self._opt_prefixes = {"-", "--"}
-        self._args: t.List[Argument] = []
+        self._args: list[_Argument] = []
 
     def add_option(
         self,
-        obj: "CoreOption",
-        opts: t.Sequence[str],
-        dest: t.Optional[str],
-        action: t.Optional[str] = None,
+        obj: CoreOption,
+        opts: cabc.Sequence[str],
+        dest: str | None,
+        action: str | None = None,
         nargs: int = 1,
-        const: t.Optional[t.Any] = None,
+        const: t.Any | None = None,
     ) -> None:
         """Adds a new option named `dest` to the parser.  The destination
         is not inferred (unlike with optparse) and needs to be explicitly
@@ -307,34 +277,32 @@ class OptionParser:
         The `obj` can be used to identify the option in the order list
         that is returned from the parser.
         """
-        opts = [normalize_opt(opt, self.ctx) for opt in opts]
-        option = Option(obj, opts, dest, action=action, nargs=nargs, const=const)
+        opts = [_normalize_opt(opt, self.ctx) for opt in opts]
+        option = _Option(obj, opts, dest, action=action, nargs=nargs, const=const)
         self._opt_prefixes.update(option.prefixes)
         for opt in option._short_opts:
             self._short_opt[opt] = option
         for opt in option._long_opts:
             self._long_opt[opt] = option
 
-    def add_argument(
-        self, obj: "CoreArgument", dest: t.Optional[str], nargs: int = 1
-    ) -> None:
+    def add_argument(self, obj: CoreArgument, dest: str | None, nargs: int = 1) -> None:
         """Adds a positional argument named `dest` to the parser.
 
         The `obj` can be used to identify the option in the order list
         that is returned from the parser.
         """
-        self._args.append(Argument(obj, dest=dest, nargs=nargs))
+        self._args.append(_Argument(obj, dest=dest, nargs=nargs))
 
     def parse_args(
-        self, args: t.List[str]
-    ) -> t.Tuple[t.Dict[str, t.Any], t.List[str], t.List["CoreParameter"]]:
+        self, args: list[str]
+    ) -> tuple[dict[str, t.Any], list[str], list[CoreParameter]]:
         """Parses positional arguments and returns ``(values, args, order)``
         for the parsed options and arguments as well as the leftover
         arguments if there are any.  The order is a list of objects as they
         appear on the command line.  If arguments appear multiple times they
         will be memorized multiple times as well.
         """
-        state = ParsingState(args)
+        state = _ParsingState(args)
         try:
             self._process_args_for_options(state)
             self._process_args_for_args(state)
@@ -343,7 +311,7 @@ class OptionParser:
                 raise
         return state.opts, state.largs, state.order
 
-    def _process_args_for_args(self, state: ParsingState) -> None:
+    def _process_args_for_args(self, state: _ParsingState) -> None:
         pargs, args = _unpack_args(
             state.largs + state.rargs, [x.nargs for x in self._args]
         )
@@ -354,7 +322,7 @@ class OptionParser:
         state.largs = args
         state.rargs = []
 
-    def _process_args_for_options(self, state: ParsingState) -> None:
+    def _process_args_for_options(self, state: _ParsingState) -> None:
         while state.rargs:
             arg = state.rargs.pop(0)
             arglen = len(arg)
@@ -391,7 +359,7 @@ class OptionParser:
         # not a very interesting subset!
 
     def _match_long_opt(
-        self, opt: str, explicit_value: t.Optional[str], state: ParsingState
+        self, opt: str, explicit_value: str | None, state: _ParsingState
     ) -> None:
         if opt not in self._long_opt:
             from difflib import get_close_matches
@@ -420,14 +388,14 @@ class OptionParser:
 
         option.process(value, state)
 
-    def _match_short_opt(self, arg: str, state: ParsingState) -> None:
+    def _match_short_opt(self, arg: str, state: _ParsingState) -> None:
         stop = False
         i = 1
         prefix = arg[0]
         unknown_options = []
 
         for ch in arg[1:]:
-            opt = normalize_opt(f"{prefix}{ch}", self.ctx)
+            opt = _normalize_opt(f"{prefix}{ch}", self.ctx)
             option = self._short_opt.get(opt)
             i += 1
 
@@ -461,7 +429,7 @@ class OptionParser:
             state.largs.append(f"{prefix}{''.join(unknown_options)}")
 
     def _get_value_from_state(
-        self, option_name: str, option: Option, state: ParsingState
+        self, option_name: str, option: _Option, state: _ParsingState
     ) -> t.Any:
         nargs = option.nargs
 
@@ -498,7 +466,7 @@ class OptionParser:
 
         return value
 
-    def _process_opts(self, arg: str, state: ParsingState) -> None:
+    def _process_opts(self, arg: str, state: _ParsingState) -> None:
         explicit_value = None
         # Long option handling happens in two parts.  The first part is
         # supporting explicitly attached values.  In any case, we will try
@@ -507,7 +475,7 @@ class OptionParser:
             long_opt, explicit_value = arg.split("=", 1)
         else:
             long_opt = arg
-        norm_long_opt = normalize_opt(long_opt, self.ctx)
+        norm_long_opt = _normalize_opt(long_opt, self.ctx)
 
         # At this point we will match the (assumed) long option through
         # the long option matching code.  Note that this allows options
@@ -529,3 +497,36 @@ class OptionParser:
                 raise
 
             state.largs.append(arg)
+
+
+def __getattr__(name: str) -> object:
+    import warnings
+
+    if name in {
+        "OptionParser",
+        "Argument",
+        "Option",
+        "split_opt",
+        "normalize_opt",
+        "ParsingState",
+    }:
+        warnings.warn(
+            f"'parser.{name}' is deprecated and will be removed in Click 9.0."
+            " The old parser is available in 'optparse'.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        return globals()[f"_{name}"]
+
+    if name == "split_arg_string":
+        from .shell_completion import split_arg_string
+
+        warnings.warn(
+            "Importing 'parser.split_arg_string' is deprecated, it will only be"
+            " available in 'shell_completion' in Click 9.0.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        return split_arg_string
+
+    raise AttributeError(name)
