@@ -1,3 +1,6 @@
+from __future__ import annotations
+
+import collections.abc as cabc
 import os
 import re
 import sys
@@ -30,10 +33,10 @@ def _posixify(name: str) -> str:
     return "-".join(name.split()).lower()
 
 
-def safecall(func: "t.Callable[P, R]") -> "t.Callable[P, t.Optional[R]]":
+def safecall(func: t.Callable[P, R]) -> t.Callable[P, R | None]:
     """Wraps a function so that it swallows exceptions."""
 
-    def wrapper(*args: "P.args", **kwargs: "P.kwargs") -> t.Optional[R]:
+    def wrapper(*args: P.args, **kwargs: P.kwargs) -> R | None:
         try:
             return func(*args, **kwargs)
         except Exception:
@@ -112,10 +115,10 @@ class LazyFile:
 
     def __init__(
         self,
-        filename: t.Union[str, "os.PathLike[str]"],
+        filename: str | os.PathLike[str],
         mode: str = "r",
-        encoding: t.Optional[str] = None,
-        errors: t.Optional[str] = "strict",
+        encoding: str | None = None,
+        errors: str | None = "strict",
         atomic: bool = False,
     ):
         self.name: str = os.fspath(filename)
@@ -123,7 +126,7 @@ class LazyFile:
         self.encoding = encoding
         self.errors = errors
         self.atomic = atomic
-        self._f: t.Optional[t.IO[t.Any]]
+        self._f: t.IO[t.Any] | None
         self.should_close: bool
 
         if self.name == "-":
@@ -175,18 +178,18 @@ class LazyFile:
         if self.should_close:
             self.close()
 
-    def __enter__(self) -> "LazyFile":
+    def __enter__(self) -> LazyFile:
         return self
 
     def __exit__(
         self,
-        exc_type: t.Optional[t.Type[BaseException]],
-        exc_value: t.Optional[BaseException],
-        tb: t.Optional[TracebackType],
+        exc_type: type[BaseException] | None,
+        exc_value: BaseException | None,
+        tb: TracebackType | None,
     ) -> None:
         self.close_intelligently()
 
-    def __iter__(self) -> t.Iterator[t.AnyStr]:
+    def __iter__(self) -> cabc.Iterator[t.AnyStr]:
         self.open()
         return iter(self._f)  # type: ignore
 
@@ -198,30 +201,30 @@ class KeepOpenFile:
     def __getattr__(self, name: str) -> t.Any:
         return getattr(self._file, name)
 
-    def __enter__(self) -> "KeepOpenFile":
+    def __enter__(self) -> KeepOpenFile:
         return self
 
     def __exit__(
         self,
-        exc_type: t.Optional[t.Type[BaseException]],
-        exc_value: t.Optional[BaseException],
-        tb: t.Optional[TracebackType],
+        exc_type: type[BaseException] | None,
+        exc_value: BaseException | None,
+        tb: TracebackType | None,
     ) -> None:
         pass
 
     def __repr__(self) -> str:
         return repr(self._file)
 
-    def __iter__(self) -> t.Iterator[t.AnyStr]:
+    def __iter__(self) -> cabc.Iterator[t.AnyStr]:
         return iter(self._file)
 
 
 def echo(
-    message: t.Optional[t.Any] = None,
-    file: t.Optional[t.IO[t.Any]] = None,
+    message: t.Any | None = None,
+    file: t.IO[t.Any] | None = None,
     nl: bool = True,
     err: bool = False,
-    color: t.Optional[bool] = None,
+    color: bool | None = None,
 ) -> None:
     """Print a message and newline to stdout or a file. This should be
     used instead of :func:`print` because it provides better support
@@ -274,7 +277,7 @@ def echo(
 
     # Convert non bytes/text into the native string type.
     if message is not None and not isinstance(message, (str, bytes, bytearray)):
-        out: t.Optional[t.Union[str, bytes]] = str(message)
+        out: str | bytes | None = str(message)
     else:
         out = message
 
@@ -319,7 +322,7 @@ def echo(
     file.flush()
 
 
-def get_binary_stream(name: "te.Literal['stdin', 'stdout', 'stderr']") -> t.BinaryIO:
+def get_binary_stream(name: t.Literal["stdin", "stdout", "stderr"]) -> t.BinaryIO:
     """Returns a system stream for byte processing.
 
     :param name: the name of the stream to open.  Valid names are ``'stdin'``,
@@ -332,9 +335,9 @@ def get_binary_stream(name: "te.Literal['stdin', 'stdout', 'stderr']") -> t.Bina
 
 
 def get_text_stream(
-    name: "te.Literal['stdin', 'stdout', 'stderr']",
-    encoding: t.Optional[str] = None,
-    errors: t.Optional[str] = "strict",
+    name: t.Literal["stdin", "stdout", "stderr"],
+    encoding: str | None = None,
+    errors: str | None = "strict",
 ) -> t.TextIO:
     """Returns a system stream for text processing.  This usually returns
     a wrapped stream around a binary stream returned from
@@ -353,10 +356,10 @@ def get_text_stream(
 
 
 def open_file(
-    filename: t.Union[str, "os.PathLike[str]"],
+    filename: str | os.PathLike[str],
     mode: str = "r",
-    encoding: t.Optional[str] = None,
-    errors: t.Optional[str] = "strict",
+    encoding: str | None = None,
+    errors: str | None = "strict",
     lazy: bool = False,
     atomic: bool = False,
 ) -> t.IO[t.Any]:
@@ -390,19 +393,19 @@ def open_file(
     """
     if lazy:
         return t.cast(
-            t.IO[t.Any], LazyFile(filename, mode, encoding, errors, atomic=atomic)
+            "t.IO[t.Any]", LazyFile(filename, mode, encoding, errors, atomic=atomic)
         )
 
     f, should_close = open_stream(filename, mode, encoding, errors, atomic=atomic)
 
     if not should_close:
-        f = t.cast(t.IO[t.Any], KeepOpenFile(f))
+        f = t.cast("t.IO[t.Any]", KeepOpenFile(f))
 
     return f
 
 
 def format_filename(
-    filename: "t.Union[str, bytes, os.PathLike[str], os.PathLike[bytes]]",
+    filename: str | bytes | os.PathLike[str] | os.PathLike[bytes],
     shorten: bool = False,
 ) -> str:
     """Format a filename as a string for display. Ensures the filename can be
@@ -518,7 +521,7 @@ class PacifyFlushWrapper:
 
 
 def _detect_program_name(
-    path: t.Optional[str] = None, _main: t.Optional[ModuleType] = None
+    path: str | None = None, _main: ModuleType | None = None
 ) -> str:
     """Determine the command used to run the program, for use in help
     text. If a file or entry point was executed, the file name is
@@ -573,12 +576,12 @@ def _detect_program_name(
 
 
 def _expand_args(
-    args: t.Iterable[str],
+    args: cabc.Iterable[str],
     *,
     user: bool = True,
     env: bool = True,
     glob_recursive: bool = True,
-) -> t.List[str]:
+) -> list[str]:
     """Simulate Unix shell expansion with Python functions.
 
     See :func:`glob.glob`, :func:`os.path.expanduser`, and
