@@ -189,7 +189,7 @@ def test_args_and_chain(runner):
     assert result.output.splitlines() == ["cli=", "a=", "b=", "c="]
 
 
-def test_multicommand_arg_behavior(runner):
+def test_group_arg_behavior(runner):
     with pytest.raises(RuntimeError):
 
         @click.group(chain=True)
@@ -219,7 +219,7 @@ def test_multicommand_arg_behavior(runner):
 
 
 @pytest.mark.xfail
-def test_multicommand_chaining(runner):
+def test_group_chaining(runner):
     @click.group(chain=True)
     def cli():
         debug()
