@@ -1,8 +1,8 @@
 import pytest
 
 import click
-from click.parser import OptionParser
-from click.parser import split_arg_string
+from click.parser import _OptionParser
+from click.shell_completion import split_arg_string
 
 
 @pytest.mark.parametrize(
@@ -20,13 +20,13 @@ def test_split_arg_string(value, expect):
 
 
 def test_parser_default_prefixes():
-    parser = OptionParser()
+    parser = _OptionParser()
     assert parser._opt_prefixes == {"-", "--"}
 
 
 def test_parser_collects_prefixes():
     ctx = click.Context(click.Command("test"))
-    parser = OptionParser(ctx)
+    parser = _OptionParser(ctx)
     click.Option("+p", is_flag=True).add_to_parser(parser, ctx)
     click.Option("!e", is_flag=True).add_to_parser(parser, ctx)
     assert parser._opt_prefixes == {"-", "--", "+", "!"}
