@@ -245,7 +245,7 @@ def test_prompt_cast_default(capfd, monkeypatch):
     monkeypatch.setattr(sys, "stdin", StringIO("\n"))
     value = click.prompt("value", default="100", type=int)
     capfd.readouterr()
-    assert type(value) is int  # noqa E721
+    assert isinstance(value, int)
 
 
 @pytest.mark.skipif(WIN, reason="Test too complex to make work windows.")
